Fix handling of Specvars
Until now the intersection of a specvar with something, yielded an error,
because it was not planned that we needed such an intersection.
But if a specvar is part of a compound, such an intersection happends

If now a specvar is intersected with something else, the result is an And-Spec.
Specvars are now correctly passed on.
add-type-to-dom can handle specvars now. If a specvar is added, we add an edge to this
specvar